refacto form and fix error window
