+++ src/app/page.tsx
@@ -3,7 +3,7 @@
 import Image from "next/image";
 
 import { NextUIProvider, Spacer } from "@nextui-org/react";
-import Navbar from "@/components/navbar";
+import Navbar from "@/components/Navbar";
 import PageSection from "@/components/section";
 
 const featuresList = [
@@ -40,7 +40,6 @@ const featuresList = [
 ];
 
 export default function Home() {
-
   return (
     <NextUIProvider>
       <>
